fix / codereview feedback

+++ cpp/include/raft/neighbors/cagra_types.hpp
@@ -317,7 +317,8 @@ struct index : ann::index {
   {
     size_t padded_dim = round_up_safe<size_t>(dataset.extent(1) * sizeof(T), 16) / sizeof(T);
 
-    if ((dataset_.extent(0) != dataset.extent(0)) || (dataset_.extent(1) != padded_dim)) {
+    if ((dataset_.extent(0) != dataset.extent(0)) ||
+        (static_cast<size_t>(dataset_.extent(1)) != padded_dim)) {
       // clear existing memory before allocating to prevent OOM errors on large datasets
       if (dataset_.size()) { dataset_ = make_device_matrix<T, int64_t>(res, 0, 0); }
       dataset_ = make_device_matrix<T, int64_t>(res, dataset.extent(0), padded_dim);
